requirements.txt
numpy>=1.18.4
pandas>=1.0.3
scikit-learn>=0.23
xgboost==0.90
neptune-client>=0.4
Flask>=1.1
Flask-WTF>=0.14.
Flask-SQLAlchemy>=2.1
psycopg2>=2.8
Flask-Compress>=1.5
matplotlib>=3.2
seaborn>=0.10
waitress>=1.4
shap>=0.35
pytz>=2020.1
tzlocal>=2.1
